feat(parser): add support for parsing entities in the grammar
The parser now supports parsing entities in the grammar. The `Package` struct now has an `entities` field that is a vector of `Entity` structs. The `Entity` struct has a `name` field that is a string. The `entity` rule in the grammar now supports optional `scope` and `attributes` fields. The `parse` function now iterates over the pairs to find the `package` rule and then iterates over the inner pairs to find the `entities` rule. The `entities` rule is then iterated over to find the `entity` rule and the `name` field is extracted from it.