feat(parser): add support for parsing entity scope and methods
This commit adds support for parsing entity scope and methods. The `Entity` struct now has a `Scope` field and a `methods` field which is a vector of `Method` structs. The `Method` struct has a `name` field. The `grammar_inc.pest` file has been updated to include the new rules for parsing entity scope and methods. The `parser/mod.rs` file has been updated to parse the new rules and populate the `Entity` struct accordingly.

Also, added some tests to ensure the new functionality works as expected.